# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WebForge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export interface Me {
  id: string;
  clerkUserId: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  imageUrl?: string | null;
  createdAt: string;
}

export type SiteAnalysisType =
  (typeof SiteAnalysisType)[keyof typeof SiteAnalysisType];

export const SiteAnalysisType = {
  website: "website",
  bot: "bot",
  backend: "backend",
  tool: "tool",
} as const;

export interface SiteAnalysis {
  type: SiteAnalysisType;
  intent: string;
  audience?: string | null;
  features: string[];
  pages: string[];
  styleHints: string[];
}

export interface SitePlanPage {
  path: string;
  title: string;
  purpose: string;
  sections: string[];
}

export type SitePlanType = (typeof SitePlanType)[keyof typeof SitePlanType];

export const SitePlanType = {
  website: "website",
  bot: "bot",
  backend: "backend",
  tool: "tool",
} as const;

export type SitePlanStyles = {
  palette: string;
  mood: string;
};

export interface SitePlan {
  type: SitePlanType;
  summary: string;
  pages: SitePlanPage[];
  styles: SitePlanStyles;
  features: string[];
  notes: string[];
}

export type SiteStatus = (typeof SiteStatus)[keyof typeof SiteStatus];

export const SiteStatus = {
  queued: "queued",
  analyzing: "analyzing",
  awaiting_confirmation: "awaiting_confirmation",
  building: "building",
  ready: "ready",
  failed: "failed",
} as const;

export type SiteCustomDomainStatus =
  | (typeof SiteCustomDomainStatus)[keyof typeof SiteCustomDomainStatus]
  | null;

export const SiteCustomDomainStatus = {
  pending: "pending",
  verified: "verified",
  failed: "failed",
} as const;

export interface Site {
  id: string;
  name: string;
  slug: string;
  prompt: string;
  status: SiteStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  message?: string | null;
  error?: string | null;
  previewUrl?: string | null;
  publicUrl?: string | null;
  coverColor?: string | null;
  files: string[];
  analysis?: SiteAnalysis | null;
  plan?: SitePlan | null;
  customDomain?: string | null;
  customDomainStatus?: SiteCustomDomainStatus;
  customDomainError?: string | null;
  customDomainTxtName?: string | null;
  customDomainTxtValue?: string | null;
  customDomainTarget?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSiteInput {
  /**
   * @minLength 4
   * @maxLength 1000
   */
  prompt: string;
  /** @maxLength 80 */
  name?: string | null;
}

export interface EditSiteInput {
  /**
   * @minLength 4
   * @maxLength 1000
   */
  prompt: string;
}

export interface SetDomainInput {
  /**
   * @minLength 3
   * @maxLength 253
   */
  domain: string;
}

export interface SendMessageInput {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  agent: "agent",
  system: "system",
} as const;

export type MessageKind = (typeof MessageKind)[keyof typeof MessageKind];

export const MessageKind = {
  text: "text",
  analysis: "analysis",
  plan: "plan",
  awaiting_confirmation: "awaiting_confirmation",
  log: "log",
  build_started: "build_started",
  build_progress: "build_progress",
  build_done: "build_done",
  build_failed: "build_failed",
} as const;

export type MessageData = { [key: string]: unknown } | null;

export interface Message {
  id: string;
  siteId: string;
  role: MessageRole;
  kind: MessageKind;
  content: string;
  data?: MessageData;
  createdAt: string;
}

export type JobKind = (typeof JobKind)[keyof typeof JobKind];

export const JobKind = {
  analyze: "analyze",
  create: "create",
  edit: "edit",
  retry: "retry",
} as const;

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  queued: "queued",
  running: "running",
  done: "done",
  failed: "failed",
} as const;

export interface Job {
  id: string;
  siteId: string;
  siteName: string;
  kind: JobKind;
  status: JobStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  message?: string | null;
  createdAt: string;
  finishedAt?: string | null;
}

export type BotStatus = (typeof BotStatus)[keyof typeof BotStatus];

export const BotStatus = {
  active: "active",
  stopped: "stopped",
  error: "error",
} as const;

export interface Bot {
  id: string;
  username?: string | null;
  displayName?: string | null;
  status: BotStatus;
  lastError?: string | null;
  tokenPreview: string;
  createdAt: string;
}

export interface HostBotInput {
  /**
   * @minLength 20
   * @maxLength 200
   */
  token: string;
}