# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WebForge API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites/{id}/domain:
    post:
      operationId: setSiteDomain
      tags: [sites]
      summary: Attach a custom domain to a site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDomainInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: removeSiteDomain
      tags: [sites]
      summary: Detach the custom domain from a site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
...
components:
  schemas:
...
    SiteAnalysis:
      type: object
      properties:
        type:
          type: string
          enum: [website, bot, backend, tool]
        intent: { type: string }
        audience: { type: string, nullable: true }
        features:
          type: array
          items: { type: string }
        pages:
          type: array
          items: { type: string }
        styleHints:
          type: array
          items: { type: string }
      required: [type, intent, features, pages, styleHints]
    SitePlanPage:
      type: object
      properties:
        path: { type: string }
        title: { type: string }
        purpose: { type: string }
        sections:
          type: array
          items: { type: string }
      required: [path, title, purpose, sections]
    SitePlan:
      type: object
      properties:
        type:
          type: string
          enum: [website, bot, backend, tool]
        summary: { type: string }
        pages:
          type: array
          items: { $ref: "#/components/schemas/SitePlanPage" }
        styles:
          type: object
          properties:
            palette: { type: string }
            mood: { type: string }
          required: [palette, mood]
        features:
          type: array
          items: { type: string }
        notes:
          type: array
          items: { type: string }
      required: [type, summary, pages, styles, features, notes]
    Site:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        prompt: { type: string }
        status:
          type: string
          enum:
            [
              queued,
              analyzing,
              awaiting_confirmation,
              building,
              ready,
              failed,
            ]
        progress: { type: integer, minimum: 0, maximum: 100 }
        message: { type: string, nullable: true }
        error: { type: string, nullable: true }
        previewUrl: { type: string, nullable: true }
        publicUrl: { type: string, nullable: true }
        coverColor: { type: string, nullable: true }
        files:
          type: array
          items: { type: string }
        analysis:
          allOf:
            - $ref: "#/components/schemas/SiteAnalysis"
          nullable: true
        plan:
          allOf:
            - $ref: "#/components/schemas/SitePlan"
          nullable: true
        customDomain: { type: string, nullable: true }
        customDomainStatus:
          type: string
          enum: [pending, verified, failed]
          nullable: true
        customDomainError: { type: string, nullable: true }
        customDomainTxtName: { type: string, nullable: true }
        customDomainTxtValue: { type: string, nullable: true }
        customDomainTarget: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        [
          id,
          name,
          slug,
          prompt,
          status,
          progress,
          files,
          createdAt,
          updatedAt,
        ]
...
    SetDomainInput:
      type: object
      properties:
        domain:
          type: string
          minLength: 3
          maxLength: 253
      required: [domain]
...