# lib/api-zod/src/generated/types/site.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WebForge API
 * OpenAPI spec version: 0.1.0
 */
import type { SiteAnalysis } from "./siteAnalysis";
import type { SiteCustomDomainStatus } from "./siteCustomDomainStatus";
import type { SitePlan } from "./sitePlan";
import type { SiteStatus } from "./siteStatus";

export interface Site {
  id: string;
  name: string;
  slug: string;
  prompt: string;
  status: SiteStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  message?: string | null;
  error?: string | null;
  previewUrl?: string | null;
  publicUrl?: string | null;
  coverColor?: string | null;
  files: string[];
  analysis?: SiteAnalysis | null;
  plan?: SitePlan | null;
  customDomain?: string | null;
  customDomainStatus?: SiteCustomDomainStatus;
  customDomainError?: string | null;
  customDomainTxtName?: string | null;
  customDomainTxtValue?: string | null;
  customDomainTarget?: string | null;
  createdAt: Date;
  updatedAt: Date;
}