# lib/api-zod/src/generated/types/siteAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WebForge API
 * OpenAPI spec version: 0.1.0
 */
import type { SiteAnalysisType } from "./siteAnalysisType";

export interface SiteAnalysis {
  type: SiteAnalysisType;
  intent: string;
  audience?: string | null;
  features: string[];
  pages: string[];
  styleHints: string[];
}