# lib/api-zod/src/generated/types/sitePlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WebForge API
 * OpenAPI spec version: 0.1.0
 */
import type { SitePlanPage } from "./sitePlanPage";
import type { SitePlanStyles } from "./sitePlanStyles";
import type { SitePlanType } from "./sitePlanType";

export interface SitePlan {
  type: SitePlanType;
  summary: string;
  pages: SitePlanPage[];
  styles: SitePlanStyles;
  features: string[];
  notes: string[];
}