# lib/api-zod/src/generated/types/sitePlanPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WebForge API
 * OpenAPI spec version: 0.1.0
 */

export interface SitePlanPage {
  path: string;
  title: string;
  purpose: string;
  sections: string[];
}